Allow users to reserve a scooter
Many updates to MapView to allow the user to select a scooter, see its
details, and reserve it.

Add a placeholder screen for when the user has  a reservation.

Rename a few files so there is consistent capitalization.

The code so far is still focused on the happy path. Error handling and
attention to detail are still needed for a final product.