Major updates to ReservedView
Replaces the temporary view that showed reservation details with a new
that shows a Map with an overlay showing the reservation status. This
will show some basic facts about the scooter reserved as well as the
current duration for the reservations.

There is a stop button overlayed on the map which currently does nothing.