Make the functions in Api more regular
They now all check the http status code. There was also a tweak to
make spacing match.

It may be worth asking why not extract the common part to a function:
I am worried this may be a per-mature abstraction. If there is a API
request that needs something slightly different, I think the common
implementation would make that harder to deal with. I also don't
foresee further changes to these functions beyond this.